Storage metadata is not lazy inited but index file itself is an optional value
